// map-ux/src/settings/tooltip/tooltip.types.ts
export enum TooltipEViewKey {
  Show = 'tooltipShow',
  //
  TriggerOn = 'tooltipTriggerOn',
  //
  Content = 'tooltipContent',
  ContentText = 'tooltipContentText',
  ContentFontSize = 'tooltipContentFontSize',
  ContentColor = 'tooltipContentColor',
  ContentBold = 'tooltipContentBold',
  ContentItalic = 'tooltipContentItalic',
  ContentLineHeight = 'tooltipContentLineHeight',
  ContentTextAlign = 'tooltipContentTextAlign',
  ContentRich = 'tooltipContentRich',
}